refactor: border glow movies into single class
